test(motion): consolidate the duplicated interactive-DOM test harness
Fix round 1, finding 4. A byte-identical 214-line installInteractiveDom
(react-dom createRoot support for interactive component tests under
Node) existed in both test/helpers/assistantPanel.test.js and
test/components/streamingReplySettled.test.js — the latter a copy
taken while building Task 5. Moved into test/lib/rendererTestHarness.js
as a shared export alongside the small findElement/findAllElements
walkers (also duplicated), and both consuming files now import it
instead of carrying their own copy. Pure extraction, no behavior
change — both files' existing suites re-verified green after the move.
